Splitting the utility with one more file
